+++ app/scripts/metamask-controller.js
@@ -99,7 +99,9 @@ module.exports = class MetamaskController {
     function logger (err, request, response) {
       if (err) return console.error(err)
       if (!request.isMetamaskInternal) {
-        console.log(`RPC (${originDomain}):`, request, '->', response)
+        if (global.METAMASK_DEBUG) {
+          console.log(`RPC (${originDomain}):`, request, '->', response)
+        }
         if (response.error) {
           console.error('Error in RPC response:\n', response.error)
         }
@@ -223,7 +225,9 @@ module.exports = class MetamaskController {
 
   // Log blocks
   processBlock (block) {
-    console.log(`BLOCK CHANGED: #${block.number.toString('hex')} 0x${block.hash.toString('hex')}`)
+    if (global.METAMASK_DEBUG) {
+      console.log(`BLOCK CHANGED: #${block.number.toString('hex')} 0x${block.hash.toString('hex')}`)
+    }
     this.verifyNetwork()
   }
 
@@ -274,7 +278,7 @@ module.exports = class MetamaskController {
 
   agreeToEthWarning (cb) {
     try {
-      this.configManager.setShouldntShowWarning(true)
+      this.configManager.setShouldntShowWarning()
       cb()
     } catch (e) {
       cb(e)