+++ ui/app/eth-store-warning.js
@@ -35,9 +35,10 @@ EthStoreWarning.prototype.render = function () {
           margin: '10px 10px 10px 10px',
         },
       },
-        `MetaMask is currently in beta -
-        exercise caution while handling
-        and storing your ether.
+        `The MetaMask team would like to
+        remind you that MetaMask is currently in beta - so
+        don't store large
+        amounts of ether in MetaMask.
         `),
 
       h('i.fa.fa-exclamation-triangle.fa-4', {